Unify create program calls
